修复编译报错. BoxApp 使用 RenderTargetBuffer 和 DepthStencilBuffer 进行渲染

+++ LearnDirectX12/Component/dx12lib/StructedBuffer.cpp
@@ -29,7 +29,7 @@ StructedBuffer::StructedBuffer(std::weak_ptr<Device> pDevice,
 		pSharedDevice->getD3DDevice(),
 		pCmdList->getD3DCommandList(),
 		pData,
-		sizeInByte
+		static_cast<uint32>(sizeInByte)
 	);
 
 	// create view
